chore: remove Expo starter boilerplate from mobile app
