+++ mobile/README.md
@@ -1,50 +1,38 @@
-# Welcome to your Expo app 👋
+# Mobile App
 
-This is an [Expo](https://expo.dev) project created with [`create-expo-app`](https://www.npmjs.com/package/create-expo-app).
+Aplicativo React Native (Expo Router) do projeto PDHC.
 
-## Get started
+## Fluxo recomendado
 
-1. Install dependencies
-
-   ```bash
-   npm install
-   ```
-
-2. Start the app
-
-   ```bash
-   npx expo start
-   ```
-
-In the output, you'll find options to open the app in a
-
-- [development build](https://docs.expo.dev/develop/development-builds/introduction/)
-- [Android emulator](https://docs.expo.dev/workflow/android-studio-emulator/)
-- [iOS simulator](https://docs.expo.dev/workflow/ios-simulator/)
-- [Expo Go](https://expo.dev/go), a limited sandbox for trying out app development with Expo
-
-You can start developing by editing the files inside the **app** directory. This project uses [file-based routing](https://docs.expo.dev/router/introduction).
-
-## Get a fresh project
-
-When you're ready, run:
+Na raiz do monorepo, use:
 
 ```bash
-npm run reset-project
+npm run dev
 ```
 
-This command will move the starter code to the **app-example** directory and create a blank **app** directory where you can start developing.
+Esse comando sobe:
 
-## Learn more
+- Supabase local
+- backend NestJS
+- mobile Expo na porta 8082
 
-To learn more about developing your project with Expo, look at the following resources:
+## Rodar apenas o mobile
 
-- [Expo documentation](https://docs.expo.dev/): Learn fundamentals, or go into advanced topics with our [guides](https://docs.expo.dev/guides).
-- [Learn Expo tutorial](https://docs.expo.dev/tutorial/introduction/): Follow a step-by-step tutorial where you'll create a project that runs on Android, iOS, and the web.
+Dentro da pasta mobile:
 
-## Join the community
+```bash
+npm install
+npm run emulator
+```
 
-Join our community of developers creating universal apps.
+Para tentar abrir automaticamente no Android Emulator:
 
-- [Expo on GitHub](https://github.com/expo/expo): View our open source platform and contribute.
-- [Discord community](https://chat.expo.dev): Chat with Expo users and ask questions.
+```bash
+npm run android:emulator
+```
+
+## Estrutura inicial
+
+O app foi limpo do boilerplate padrão do create-expo-app e começa com uma tela base em:
+
+- app/index.tsx
