+++ mobile/app/index.tsx
@@ -1,0 +1,31 @@
+import { StyleSheet, Text, View } from 'react-native';
+
+export default function HomeScreen() {
+  return (
+    <View style={styles.container}>
+      <Text style={styles.title}>Hospital Dom Helder Câmara</Text>
+      <Text style={styles.subtitle}>Base mobile inicial pronta para desenvolvimento.</Text>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+    padding: 24,
+    backgroundColor: '#ffffff',
+  },
+  title: {
+    fontSize: 28,
+    fontWeight: '700',
+    color: '#11181c',
+    marginBottom: 8,
+  },
+  subtitle: {
+    fontSize: 16,
+    color: '#687076',
+    textAlign: 'center',
+  },
+});
