+++ mobile/package.json
@@ -4,14 +4,13 @@
   "version": "1.0.0",
   "scripts": {
     "start": "expo start --port 8082",
-    "reset-project": "node ./scripts/reset-project.js",
     "emulator": "expo start --localhost --port 8082",
     "android": "expo start --android --port 8082",
     "android:emulator": "expo start --android --localhost --port 8082",
     "ios": "expo start --ios --port 8082",
     "web": "expo start --web --port 8082",
-    "lint": "expo lint",
-    "lint:fix": "expo lint --fix",
+    "lint": "expo lint app",
+    "lint:fix": "expo lint app -- --fix",
     "typecheck": "tsc --noEmit",
     "format": "prettier --write \"**/*.{js,jsx,ts,tsx,json,md}\"",
     "format:check": "prettier --check \"**/*.{js,jsx,ts,tsx,json,md}\""
